Renamed preview + form folders to match form component names
