+++ src/components/Main/inputForm/Checkboxes.tsx
